requirements.txt
django
djangorestframework
django-extensions
django-cors-headers
pyodbc
mssql-django